refactor: API 관련 폴더 구조 변경
- 공통 단위 (shared)
  1. client: axios 인스턴스는 서버와의 통신에 공통적으로 사용되므로 shared 폴더로 이동합니다.

그 외 모든 API는 이번 리팩토링 과정 중 분리된 entities 폴더 등 사용하는 기능에 맞는 폴더의 api 폴더로 이동했습니다.
