+++ src/shared/api/index.js
@@ -1,0 +1,1 @@
+export { client } from "./client";
