+++ src/widgets/sign-in-form/ui/SignInForm.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
-import { useForm } from "react-hook-form";
-import { signIn } from "../../../api/user";
-import { useNavigate } from "react-router-dom";
 import styles from "../styles.module.css";
+import { useForm } from "react-hook-form";
+import { useNavigate } from "react-router-dom";
+import { signIn } from "../api";
 
 export default function SignInForm() {
   const {
